added displayed cells, still need to fix tho.

+++ Assignment16/Assignment16/Models/PlanetData.swift
@@ -10,5 +10,5 @@ import UIKit
 struct PlanetData {
     let image: UIImage
     let name: String
-    let surfaceArea: Double
+    let surfaceArea: Int
 }
